Add CSS styling to match design
Add a lot of styling to match design as close as I can.
The spacing could definitely be a bit better but I honestly
think everything is pretty close. I am not very happy with it
though because the footer on my bigger monitor floats with
white space beneath it, so that needs to be fixed. Ideally
the footer would be at the bottom on my regular monitor without
needing to scroll, and everything would just shrink to fit,
so those will be my next steps.